Crashing + slow, gotta check with the existing num of allocated DSes first.

+++ idk/projects/vkn/DeferredPass.cpp
@@ -431,7 +431,7 @@ namespace idk::vkn
 			if (oidx)
 			{
 				cmd_buffer.bindIndexBuffer(*(*oidx).buffer(), 0, mesh.IndexType(), vk::DispatchLoaderDefault{});
-				cmd_buffer.drawIndexed(mesh.IndexCount(), p_ro.num_instances, 0, 0, p_ro.inst_offset, vk::DispatchLoaderDefault{});
+				cmd_buffer.drawIndexed(mesh.IndexCount(), static_cast<uint32_t>(p_ro.num_instances), 0, 0, static_cast<uint32_t>(p_ro.inst_offset), vk::DispatchLoaderDefault{});
 			}
 		}
 		return rendered;
